Allow building with Java 17
Add a new 'java17' Maven build profile which is auto-activated when
Java 17 is detected.

N.B. Three speed characterization profiles are currently being excluded
from compilation on Java 17 due to breaking API changes in
datasketches-memory-4.x.
This was done to avoid needing to introduce separate source folders and
compilation units to handle compiling across datasketches-memory-3.x
and datasketches-memory-4.x with Java 8/11 and 17 respectively.
This should allow Maven builds to execute cleanly and opens the door for
adding GitHub Action workflows/CI in the future.
It's not clear if this allows for a seamless local developer workflow with
miscellaneous IDEs. (Eclipse, IntelliJ IDEA)
Other possibilities for handling these source-level incompatibilities include:
  - Preserving APIs required by datasketches-charaterization across the
     datasketches library versions actively supported/tested
  - Introducing separate source directories/submodules as per Maven docs:
     https://maven.apache.org/plugins/maven-compiler-plugin/multirelease.html#patterns-summary